+++ fabric/src/main/java/de/cech12/bucketlib/platform/FabricFluidHelper.java
@@ -167,7 +167,7 @@ public class FabricFluidHelper implements IFluidHelper {
                 transaction.commit();
             }
             ItemStack resultStack = context.getItemVariant().toStack();
-            if (!resultStack.isEmpty()) {
+            if (!resultStack.isEmpty() && FluidStorageData.EMPTY.equals(resultStack.get(BucketLibMod.STORAGE))) {
                 resultStack.remove(BucketLibMod.STORAGE);
             }
             return resultStack;