+++ neoforge/src/main/java/de/cech12/bucketlib/platform/NeoforgePlatformHelper.java
@@ -31,7 +31,7 @@ public class NeoforgePlatformHelper implements IPlatformHelper {
 
     @Override
     public boolean isDevelopmentEnvironment() {
-        return !FMLLoader.isProduction();
+        return !FMLLoader.getCurrent().isProduction();
     }
 
     @Override