+++ neoforge/src/main/java/de/cech12/bucketlib/rei/BucketLibReiClientPlugin.java
@@ -4,6 +4,7 @@ import de.cech12.bucketlib.api.BucketLibTags;
 import de.cech12.bucketlib.api.item.UniversalBucketItem;
 import de.cech12.bucketlib.platform.Services;
 import de.cech12.bucketlib.util.BucketLibUtil;
+import de.cech12.bucketlib.util.RegistryUtil;
 import me.shedaniel.rei.api.client.plugins.REIClientPlugin;
 import me.shedaniel.rei.api.client.registry.display.DisplayRegistry;
 import me.shedaniel.rei.api.common.util.EntryIngredients;
@@ -12,7 +13,6 @@ import me.shedaniel.rei.plugin.common.displays.DefaultFuelDisplay;
 import me.shedaniel.rei.plugin.common.displays.anvil.DefaultAnvilDisplay;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.data.registries.VanillaRegistries;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.EnchantmentInstance;
@@ -23,6 +23,7 @@ import net.minecraft.world.level.material.Fluids;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.OptionalInt;
 
 @SuppressWarnings("unused")
 @REIPluginClient
@@ -47,7 +48,7 @@ public class BucketLibReiClientPlugin implements REIClientPlugin {
         }
         //register infinity enchanting
         if (Services.CONFIG.isInfinityEnchantmentEnabled()) {
-            EnchantmentInstance data = new EnchantmentInstance(VanillaRegistries.createLookup().lookup(Registries.ENCHANTMENT).get().getOrThrow(Enchantments.INFINITY), 1);
+            EnchantmentInstance data = new EnchantmentInstance(RegistryUtil.getRegistryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(Enchantments.INFINITY), 1);
             for (UniversalBucketItem bucketItem : Services.REGISTRY.getRegisteredBuckets()) {
                 for (Fluid fluid : Services.REGISTRY.getAllFluids()) {
                     if (fluid != Fluids.EMPTY && bucketItem.canHoldFluid(fluid) && fluid.defaultFluidState().is(BucketLibTags.Fluids.INFINITY_ENCHANTABLE)) {
@@ -57,7 +58,8 @@ public class BucketLibReiClientPlugin implements REIClientPlugin {
                         registry.add(new DefaultAnvilDisplay(
                                 List.of(EntryIngredients.of(bucket), EntryIngredients.of(EnchantmentHelper.createBook(data))),
                                 List.of(EntryIngredients.of(enchantedBucket)),
-                                Optional.empty()));
+                                Optional.empty()),
+                                OptionalInt.of(4));
                     }
                 }
             }