+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [1.21.9-4.6.0.0] - 2026-04-22
+### Changed
+- Updated to Minecraft 1.21.9 (Fabric 0.134.1+1.21.9, NeoForge 21.9.16-beta)
+- Updated Cloth Config support (20.0.149) (Fabric/Quilt)
+- Updated ModMenu support (16.0.1) (Fabric/Quilt)
+- Updated JEI support (25.0.1.4) (NeoForge)
+- Updated REI support (21.9.813) (NeoForge, Fabric/Quilt)
+
 ## [1.21.7-4.5.0.1] - 2026-04-18
 ### Changed
 - re-added JEI as optional dependency for NeoForge